+++ packages/snap-logger/docs/assets/js/main.js
@@ -204,31 +204,37 @@
 			},
 
 		/******/
-	}; // The module cache
+	};
 	/************************************************************************/
-	/******/ /******/ var __webpack_module_cache__ = {}; // The require function
+	/******/ // The module cache
+	/******/ var __webpack_module_cache__ = {};
 	/******/
-	/******/ /******/ function __webpack_require__(moduleId) {
+	/******/ // The require function
+	/******/ function __webpack_require__(moduleId) {
 		/******/ // Check if module is in cache
 		/******/ if (__webpack_module_cache__[moduleId]) {
 			/******/ return __webpack_module_cache__[moduleId].exports;
 			/******/
-		} // Create a new module (and put it into the cache)
-		/******/ /******/ var module = (__webpack_module_cache__[moduleId] = {
+		}
+		/******/ // Create a new module (and put it into the cache)
+		/******/ var module = (__webpack_module_cache__[moduleId] = {
 			/******/ // no module.id needed
 			/******/ // no module.loaded needed
 			/******/ exports: {},
 			/******/
-		}); // Execute the module function
+		});
 		/******/
-		/******/ /******/ __webpack_modules__[moduleId](module, module.exports, __webpack_require__); // Return the exports of the module
+		/******/ // Execute the module function
+		/******/ __webpack_modules__[moduleId](module, module.exports, __webpack_require__);
 		/******/
-		/******/ /******/ return module.exports;
+		/******/ // Return the exports of the module
+		/******/ return module.exports;
 		/******/
-	} /* webpack/runtime/compat get default export */
+	}
 	/******/
 	/************************************************************************/
-	/******/ /******/ (() => {
+	/******/ /* webpack/runtime/compat get default export */
+	/******/ (() => {
 		/******/ // getDefaultExport function for compatibility with non-harmony modules
 		/******/ __webpack_require__.n = (module) => {
 			/******/ var getter = module && module.__esModule ? /******/ () => module['default'] : /******/ () => module;
@@ -237,9 +243,10 @@
 			/******/
 		};
 		/******/
-	})(); /* webpack/runtime/define property getters */
+	})();
 	/******/
-	/******/ /******/ (() => {
+	/******/ /* webpack/runtime/define property getters */
+	/******/ (() => {
 		/******/ // define getter functions for harmony exports
 		/******/ __webpack_require__.d = (exports, definition) => {
 			/******/ for (var key in definition) {
@@ -252,14 +259,16 @@
 			/******/
 		};
 		/******/
-	})(); /* webpack/runtime/hasOwnProperty shorthand */
+	})();
 	/******/
-	/******/ /******/ (() => {
+	/******/ /* webpack/runtime/hasOwnProperty shorthand */
+	/******/ (() => {
 		/******/ __webpack_require__.o = (obj, prop) => Object.prototype.hasOwnProperty.call(obj, prop);
 		/******/
-	})(); /* webpack/runtime/make namespace object */
+	})();
 	/******/
-	/******/ /******/ (() => {
+	/******/ /* webpack/runtime/make namespace object */
+	/******/ (() => {
 		/******/ // define __esModule on exports
 		/******/ __webpack_require__.r = (exports) => {
 			/******/ if (typeof Symbol !== 'undefined' && Symbol.toStringTag) {
@@ -270,10 +279,12 @@
 			/******/
 		};
 		/******/
-	})(); // startup // Load entry module
+	})();
 	/******/
 	/************************************************************************/
-	/******/ /******/ /******/ __webpack_require__('./default/assets/js/src/bootstrap.ts');
+	/******/ // startup
+	/******/ // Load entry module
+	/******/ __webpack_require__('./default/assets/js/src/bootstrap.ts');
 	/******/ // This entry module used 'exports' so it can't be inlined
 	/******/
 })();