+++ packages/snap-client/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.3.5 (2021-06-21)
+
+**Note:** Version bump only for package @searchspring/snap-client
+
+
+
+
+
+## 0.3.4 (2021-06-18)
+
+**Note:** Version bump only for package @searchspring/snap-client
+
+
+
+
+
+## 0.3.3 (2021-06-18)
+
+**Note:** Version bump only for package @searchspring/snap-client
+
+
+
+
+
+## 0.3.2 (2021-06-18)
+
+**Note:** Version bump only for package @searchspring/snap-client
+
+
+
+
+
+## 0.3.1 (2021-06-18)
+
+**Note:** Version bump only for package @searchspring/snap-client
+
+
+
+
+
+# 0.3.0 (2021-06-17)
+
+
+### Features
+
+* **slideout.tsx:** slideout slide direction ([#45](https://github.com/searchspring/snap/issues/45)) ([55b6544](https://github.com/searchspring/snap/commit/55b654489d66c48cea1fe4aa6769c14fcd779a4f))
+
+
+
+
+
+## 0.2.1 (2021-06-15)
+
+**Note:** Version bump only for package @searchspring/snap-client
+
+
+
+
+
 # 0.2.0 (2021-05-20)